Handling recv msgs, pub/sub client in one class
Added basic handling of received messages by the server. Checking for
the type of received message and calling the proper function from the
PubSubHub class.

Also combined the seperate publisher and subscriber classes into one
PubSubClient class. This will allow a client to be both publisher and
subscriber reducing the amount of sockets required.